requirements.txt
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
tzdata==2025.2
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==21.2.0
